Add DirectInputs for reading inputs directly from a DataFrame.
